requirements.txt
Django==2.2.6
django-markdownx==2.0.28
Markdown==3.1.1
Pillow==6.2.0
pytz==2019.3
sqlparse==0.3.0
django-widget-tweaks==1.4.5
django-filter==2.2.0